Added fixed point update_local_code
This is more efficient than using floating point rounding and more
accurate than relying on the sampling rate.

Note there was a bug in the original code wherein the early and late
local replicas are actually the late and early respectively. I've
continued this for now, but it should be fixed.