Improvements to gnss_flowgraph
1) Fixed a bug in gnss_signal::compare

The signal string comparison was not correct.

2) Fix for gnss_flowgraph when satellites are specified

3) Better handling of acquisition in flow graph

Now PRNs are alternated amongst channels, so we don't get 'stuck' on a
particular PRN, and we don't miss out on PRNs that are assigned to
channels that come after a channel that is always in the acquisition state.

4) Added SYS.prns to config file.

Now we can specify which PRNs are available in the configuration, rather
than re-compiling. Particularly useful for debugging, or experimenting
with files where you already know which signals are present