Bugfix for Spirent PRS code generator
Not sure about this one. From my understanding of the Spirent knowledge
base description of the way in which the PRS code is generated, the
first PRS code chip for PRN 1 should be the same as the second p-code
chip for PRN 1, but it appears that in fact what is generated is the
last chip of the week. In other words the offset is -1 not +1.

With this modification I can successfully process E1 PRS data for both
even and odd numbered PRNs