Added message port observables->channel
This allows the observables to send messages back to the tracking
channels, for example when system time has been received by at least one
channel, it can then be broadcast back to all other channels.

This is useful in long code tracking, so that we don't have to receive a
TOW on every single channel before we start tracking.