Added accumulate_array function
This is a utility function similar to Matlab/Octave accumarray. It
accumulates values from an input array, into an output array according
to indices provided by an index array. The input array and the index
array are the same length, the output array length must be sufficient
such that the largest element of the index array is less than the length
of the output array. Thus for example, output_array[0] will contain the
sum of all components of the input_array for which the corresponding
elements in the index_array are 0.