Added codess EPL and VEPL correlators
This is still experimental. The idea is that the signal is coherently
accumulated over 1 code chip [with subcarrier wipeoff in the case of
VEPL processing]. The array of individual code chip partial integrals
are then squared (complex square, not square magnitude) and summed.

Note that this:
1) Greatly reduces the processing gain
2) Provides no cross correlation protection, except for frequency offset
and subcarrier orthogonality [in the case of VEPL]