Improve portability
In some architectures (e.g. alpha, hppa, powerpcspe, m68k,  sh4, sparc64, x32) the package gr-osmosdr is not available. So when the package is build with -DENABLE_OSMOSDR=ON, it breaks on the mentioned architectures. This is expected behaviour (it breaks because a required dependency is not found), but prevents from building the package on such architectures. This commit introduces a small change: when compilation is called with

cmake -DENABLE_OSMOSDR=ON -DENABLE_PACKAGING=ON ..

then, the compilation does not break if gr-osmosdr is not found.

This commit also fixes building when UHD is present but gnuradio-uhd is not (for instance, in hurd-i386)