Fix register unpacking for Labsat3W files
This change fixes a bug in the unpacking of Labsat 3 Wideband files
when using the Labsat_Signal_Source. The original endian conversion
loop includes a cast from char->uint64_t that (surprisingly) incurs a sign
extension when the MSB of the char is set. ORing the unmasked uint64_t
into the register can set undesired bits.

The changes replace the old endian conversion loop with a ``boost`` utility
function.

Signed-off-by: Jeff Melville <dev@jeffmelville.com>