Removing a TODO
Answer:
No, the default constructor for `shared_ptr<T>` does not instantiate the contained type, it starts with a nullptr. This is the reason for `std::make_shared<T>()`.
You can also use the constructor if you want to manage an already existing pointer by passing it as an argument (ie: `std::shared_ptr<T>{new T{}}`).