Implement 32f sincos with RVV
Implement experimental 32-bit floating-point number generation
of complex numbers using sin and cos utilizing RVV C intrinsics.
To be fully honest, simply reverse-engineered the NEON implementation,
and not entirely sure how the final two big calculations ("extended
precision modular arithmetic" and calculating both polynomials)
were derived. As such, it is possible there is a method that better
utilizes the capabilities of RVV, and welcome any change that would
be able to take advantage of it.

Signed-off-by: Marcus Alagar <mvala079@gmail.com>